updated database.c and interprocess_communication.h

+++ server/src/database.c
@@ -19,14 +19,21 @@ void create_tables()
                   "    rcpt_username VARCHAR(50), "
                   "    message VARCHAR(2048) "
                   ");";
-    PQexec(conn, query);
+    PGresult* res = PQexec(conn, query);
+
+    if (PQresultStatus(res) != PGRES_COMMAND_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+    }
+
+    PQclear(res);
+    PQfinish(conn);
 
     return;
 }
 
 void add_new_user(char *username, char *password)
 {
-    PGconn *conn = PQconnectdb("host=<host> port=<port> dbname=<dbname> user=<user> password=<password>");
+    PGconn *conn = PQconnectdb("host=localhost port=5432 dbname=postgres user=postgres password=postgres");
 
     if (PQstatus(conn) != CONNECTION_OK) {
         fprintf(stderr, "Connection to database failed: %s", PQerrorMessage(conn));
@@ -37,14 +44,21 @@ void add_new_user(char *username, char *password)
     char query[MAX_QUERY_SIZE];
     sprintf(query, "INSERT INTO users "
                    "VALUES (%s, %s); ", username, password);
-    PQexec(conn, query);
+    PGresult* res = PQexec(conn, query);
+
+    if (PQresultStatus(res) != PGRES_COMMAND_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+    }
+
+    PQclear(res);
+    PQfinish(conn);
 
     return;
 }
 
 int find_username_from_database(char *key, char *username)
 {
-    PGconn *conn = PQconnectdb("host=<host> port=<port> dbname=<dbname> user=<user> password=<password>");
+    PGconn *conn = PQconnectdb("host=localhost port=5432 dbname=postgres user=postgres password=postgres");
 
     if (PQstatus(conn) != CONNECTION_OK) {
         fprintf(stderr, "Connection to database failed: %s", PQerrorMessage(conn));
@@ -58,19 +72,27 @@ int find_username_from_database(char *key, char *username)
                    "WHERE username = '%s' ", username);
     PGresult *res = PQexec(conn, query);
 
-    if (PQntuples(res) > 0)
-    {
-        return 0;
+    if (PQresultStatus(res) != PGRES_TUPLES_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+        PQclear(res);
+        PQfinish(conn);
+        return -1;
     }
-    else
-    {
+
+    if (PQntuples(res) > 0){
+        PQclear(res);
+        PQfinish(conn);
+        return 0;
+    } else {
+        PQclear(res);
+        PQfinish(conn);
         return 1;
     }
 }
 
 int authenticate_user_from_database(char *username, char *password)
 {
-    PGconn *conn = PQconnectdb("host=<host> port=<port> dbname=<dbname> user=<user> password=<password>");
+    PGconn *conn = PQconnectdb("host=localhost port=5432 dbname=postgres user=postgres password=postgres");
 
     if (PQstatus(conn) != CONNECTION_OK) {
         fprintf(stderr, "Connection to database failed: %s", PQerrorMessage(conn));
@@ -84,19 +106,27 @@ int authenticate_user_from_database(char *username, char *password)
                    "WHERE username = '%s' ", username);
     PGresult *res = PQexec(conn, query);
 
-    if ( strcmp(PQgetvalue(res, 0, 0), password) == 0 )
-    {
-        return 0;
+    if (PQresultStatus(res) != PGRES_TUPLES_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+        PQclear(res);
+        PQfinish(conn);
+        return -1;
     }
-    else
-    {
+
+    if ( strcmp(PQgetvalue(res, 0, 0), password) == 0 ) {
+        PQclear(res);
+        PQfinish(conn);
+        return 0;
+    } else {
+        PQclear(res);
+        PQfinish(conn);
         return 1;
     }
 }
 
 PGresult *get_buffer_messages(char *client_username, char *rcpt_username)
 {
-    PGconn *conn = PQconnectdb("host=<host> port=<port> dbname=<dbname> user=<user> password=<password>");
+    PGconn *conn = PQconnectdb("host=localhost port=5432 dbname=postgres user=postgres password=postgres");
 
     if (PQstatus(conn) != CONNECTION_OK) {
         fprintf(stderr, "Connection to database failed: %s", PQerrorMessage(conn));
@@ -111,16 +141,28 @@ PGresult *get_buffer_messages(char *client_username, char *rcpt_username)
                    "ORDER BY rcpt_username ", client_username);
     PGresult *res = PQexec(conn, query);
 
+    if (PQresultStatus(res) != PGRES_TUPLES_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+        PQclear(res);
+        PQfinish(conn);
+        return NULL;
+    }
+
     sprintf(query, "DELETE FROM buffer_messages "
                    "WHERE username = '%s' ", client_username);
-    PQexec(conn, query);
+    PGresult *res2 = PQexec(conn, query);
+
+    if (PQresultStatus(res2) != PGRES_COMMAND_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+        return NULL;
+    }
 
     return res;
 }
 
 void add_buffer_message(char *client_username, char *rcpt_username, char *message)
 {
-    PGconn *conn = PQconnectdb("host=<host> port=<port> dbname=<dbname> user=<user> password=<password>");
+    PGconn *conn = PQconnectdb("host=localhost port=5432 dbname=postgres user=postgres password=postgres");
 
     if (PQstatus(conn) != CONNECTION_OK) {
         fprintf(stderr, "Connection to database failed: %s", PQerrorMessage(conn));
@@ -130,8 +172,12 @@ void add_buffer_message(char *client_username, char *rcpt_username, char *messag
 
     char query[MAX_QUERY_SIZE];
     sprintf(query,"INSERT INTO buffer_messages "
-                   "VALUES ('%s', '%s', '%s')", client_username, rcpt_username, message);
-    PQexec(conn, query);
+                   "VALUES ('%s', '%s', '%s') ", client_username, rcpt_username, message);
+    PGresult *res2 = PQexec(conn, query);
+
+    if (PQresultStatus(res2) != PGRES_COMMAND_OK) {
+        printf("Query execution failed: %s\n", PQerrorMessage(conn));
+    }
 
     return;
 }
